% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harvest}
\alias{harvest}
\title{Sage Grouse Harvest Data by Working Group for Wyoming}
\format{A data.frame:
\describe{
  \item{Year}{The year as an integer.}
  \item{WorkingGroup}{The working group as a factor.}
  \item{Harvest}{The number of sage grouse harvested as an integer.}
  \item{HunterDays}{The number of hunter days as an integer.}
  \item{ChickWings}{The number of hunter harvested wings from chicks as an integer.}
  \item{TotalWings}{The total number of hunter harvested wings as an integer.}
}}
\usage{
harvest
}
\description{
Sage Grouse Harvest Data by Working Group for Wyoming
}
\keyword{datasets}

