% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lek}
\alias{lek}
\title{Sage Grouse Lek Observations for Wyoming}
\format{An object of class \code{data.frame} with 119703 rows and 15 columns.}
\usage{
lek
}
\description{
Greater sage grouse lek observations for
the State of Wyoming from 1948 to 2015.
}
\details{
At the request of the State of Wyoming
the geolocations of individual leks and the names of individuals
conducting lek counts are not included although
the official state individual lek identifiers are provided.
}
\keyword{datasets}
