% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lek}
\alias{lek}
\title{Sage Grouse Lek Observations for Wyoming}
\format{\preformatted{'data.frame':	119703 obs. of  15 variables:
 $ LekID       : Factor w/ 2415 levels "A-3 Bar H Road",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ Year        : int  1948 1949 1950 1951 2006 2007 2007 2008 2009 2010 ...
 $ Month       : int  NA NA NA NA NA 3 4 5 NA NA ...
 $ Day         : int  NA NA NA NA NA 29 10 14 NA NA ...
 $ Time        : Factor w/ 914 levels "","-530","?",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ Check.      : Factor w/ 2 levels "No","Yes": 2 2 2 2 1 2 2 2 1 1 ...
 $ Status      : Factor w/ 4 levels "","Active","Inactive",..: 2 2 2 2 4 4 4 4 4 4 ...
 $ Male        : int  13 18 9 7 NA 0 0 0 NA NA ...
 $ Fem         : int  0 0 0 0 NA 0 0 0 NA NA ...
 $ Unk         : int  0 0 0 0 NA 0 0 0 NA NA ...
 $ Method      : Factor w/ 4 levels "","Air","Ground",..: 3 3 3 3 1 3 3 3 1 1 ...
 $ Type        : Factor w/ 5 levels "","Count","Not Checked",..: 2 2 2 2 3 4 4 4 3 3 ...
 $ Moved       : Factor w/ 3 levels "","No","Yes": 2 2 2 2 2 2 2 2 2 2 ...
 $ Sign        : Factor w/ 5 levels "","Checked-Sign Found",..: 4 4 4 4 4 4 4 4 4 4 ...
 $ WorkingGroup: Factor w/ 8 levels "Bates Hole","Bighorn",..: 7 7 7 7 7 7 7 7 7 7 ...
}}
\usage{
lek
}
\description{
Greater sage grouse lek observations for
the State of Wyoming from 1948 to 2015.
}
\details{
At the request of the State of Wyoming
the geolocations of individual leks and the names of individuals
conducting lek counts are not included although
the official state individual lek identifiers are provided.
}
\keyword{datasets}

